library/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.18.0
Pillow>=8.3.0
matplotlib>=3.4.0
pandas>=1.3.0
seaborn>=0.11.0
jupyter>=1.0.0
napari>=0.4.0
tqdm>=4.62.0
opencv-python>=4.5.0
plotly>=5.0.0
ipywidgets>=7.6.0 
PyOpenGL-accelerate>=3.1.0